Fixed OneFichierCom.py from userplugins.
OneFichierCom.py from userplugins in home directory was failing at downloading multiple files.
When the first file was downloaded the plugin was'nt handling the waiting time.
This was due tu the fact that the sentence in the web page saying to wait has changed.
I just modified the sentence to make the plugin wait correctly before or between each download.