Movie2kTo: fixed hoster recognition
It appeared to me that the hoster and the id of a video can be present in
either a JavaScript section or in a HTML section on *one* page.
I first assumed that all hosters on one page can be found ether in a JavaScript
section or in a HTML section.