Update Captcha9kw.py
Reason for "Selfsolve" only for say to the users "activate selfsolve (9kw) in the tool pyload".

And for 300 default timeout, from 9kw.eu FAQ, quote:
"What happens when there is no user present?
In that case several possibilities are provided. After a maximum of 60 seconds the alternative system takes over and has another 30 seconds to find a solution in a different way. Optimally that should take 5-30 seconds. Each user has a maximum of 30 seconds to a captcha to solve up to maximum 300 seconds for up to 5 users. Only for captchas with timeout 300 or higher. "