Unrar 5.x Support
With Unrar 5.x the ExtractArchive plugin doesn't work for password protected archives. The reason for this is that the output of Unrar has changed. Therefore, the script must use a different regex pattern to parse the received data.

This new version of the plugin does essentially that after determining whether Unrar 5.x is used.