Fix cookie issues
The Account-Page didn't show any Values on 'Traffic left' and 'Valid
until' for my Premium Account. Therefore no premium download was
possible. The Problem occured on ArchLinux and FreeBSD, but somehow not
on Ubuntu.
The Trick is to set the value for 'expires' as type int when it is just
a number.