[filepup.net] New hoster
Once again a plugin for filepup.net (https://github.com/pyload/pyload/issues/282). The website has been changed since my first pullrequest. I wasn't able to implement the offline check because the string which says the file is offline is always present in the code, even if the file is online. There are no captchas or wait penalities with 100MB files.