Made the bug fix non-specific.
The original modification did not account for folder_per_package being set to no, causing folder parameters in that scenario to be set to "". As the introduction of the parameter includes a default value of "", the else clause which used to do this and now causes issues can simply be deleted to accommodate the change. Thanks go to Firefly for noting the bug and its simple solution.