Cosmetic Polish of PyPackage fix.
Removed the inline description of the modification and an unintended addition of leading spaces. Functionally identical: the deletion of save_path from the folder property, to prevent the bug which caused download locations, added via the API, to be mangled from X:\Intended\Folder to D:\Default pyLoad Download Folder\XintendedFolder, by stripping special characters from the path. The intent of the original code is unclear, but may have been a hold-over.