Update DlProtectCom.py
Hello

Frist, I'm a French guy so sorry for my English.

 I have made some fix for the DlProtect crypter :
- Remove '>' at the start of string to search in the html page
- Chage field name on form
- Add a loop to wait 'Continue' button on the form.

I stay 1 bug, because for me the message was in French in the page ('Please click on continue to see the content' => 'Cliquez sur continuer pour voir le(s) lien(s)'). I don't not how to manage multi-language in the crypter.