New Crypter plugin: pasted.co
Pasted.co is a pastebin service. It can be used to host a list of
download links which pyLoad could use. The plugin bypasses the
(ineffective) captcha request and fetches the contents of the paste. It
keeps the original package's name and folder and just inserts the links
from the paste.