Update ShSt crypter plugin
The plugin doesn't rely on a third party website anymore for the actual
decrypting. Instead it uses a different user agent which makes sh.st
serve an HTTP redirect. This speeds up the decrypting process and fixes
errors related to the third party web site (in particular the latest
outage and API change).
Also, self.urls is now used instead of self.packages for the decrypted
URL.