Added plugin Webinterface Notify. This Plug-in will provide Popup notifications for latest chrome and opera browsers.
The Webinterface has been modified such, that plugins can register additional content which is embedded as base64 string.
This way, each plugin still consists of only one file. this helps to keep the update procedures intact.
Files served are parsed by the template system. if the name of a file registered by a plugin contains the string "static" it won't be parsed.
This is important for binary data like images.