Shareonline.Biz: Fail on full instead of wait
Up to now, if a file cannot be downloaded because of "Server is full", it waits 10 minutes a couple of times. And while waiting, nothing else gets downloaded from Shareonline.biz. After a couple of retries, the download will fail anyway. The "server is full" message usually stays for around one day on a file, but it doesn't necessarily affect other files on same hoster.

Proposed solution: Do not wait on "server is full" but fail instead. Then, other files on ShareOnline.biz will continue downloading.