Fix tmpdir parameter for CLI start command
CLI `start` command was expecting a `tempdir` parameter, but it was
receiving `tmpdir` instead. As the command flag is named `tmpdir`, it
was changed to this name, so there's no need to change the CLI
signature.

Relevant stacktrace:
```
Traceback (most recent call last):
  File "$HOME/.virtualenvs/pyload2/bin/pyload", line 11, in <module>
    load_entry_point('pyload.core==0.6.2', 'console_scripts', 'pyload')()
  File "build/bdist.linux-x86_64/egg/pyload/core/cli.py", line 156, in main
TypeError: start() got an unexpected keyword argument 'tmpdir'
```