Avoid name mangling on pyload core object
Making the pyload core object a private (double-underscode) attribute
`__pyload` is not a good idea because of name mangling when dealing with
class inheritance.

This means the private attribute is prepended with the base class' name,
which means that, e.g. FileManager doesn't found its `__pyload`
attribute because it's stored as `_BaseManager__pyload` instead.

For more information, see:
https://docs.python.org/2/tutorial/classes.html#private-variables-and-class-local-references

This could be fixed by using the `get_core` method instead, but I
implemented a `pyload_core` property instead, to maintain the
attribute's privacy and remove the getter.

Relevant log lines:
```
2017-07-27 08:22:54  CRITICAL  'FileManager' object has no attribute '_FileManager__pyload'
```