Fix makefile when file exists in current directory
`fs.makefile` was checking the path for the file to create exists, and
then it tried to create the file. However, it tried to create it by
using just the file name, instead of the absolute path to the file. For
this reason, if a file with that name already existed in the pyload
directory, if failed and the file was not being created.

Relevant log lines:
```
2017-07-27 20:07:09  CRITICAL  Path already exists
```