Add default to storage_folder config option
This adds the current folder as the default storage location for pyLoad.
It could be changed to a better default value, but we need a default
for pyLoad to work, because of the way `availspace` is being used, as it
expects a valid path.

Failing line:
https://github.com/pyload/pyload/blob/ac38e1f0ccdb04ae3a697b83954722753a6c5b7c/src/pyload/core/manager/transfer.py#L157

Relevant log error:
```
CRITICAL  [Errno 2] No existe el archivo o el directorio: ''
```