Create MAGNET.py
For use with ReadDebridTorrent and magnet links (or any other plugin that accepts magnet links).  Useful when used in conjunction with Radarr or Sonarr or similar that can generate .magnet file links in a blackhole folder.  Only minimally tested and doesn't really contain any error checking, but tested and works well