[core] consider offline links as finished
This change considers (permanently) offline links as finished. Although
they are not technically finished in being downloaded, they also never
will be so there is no reason not to dispatch `package_finished`. This
allows plugins and scripts to further process the files.

Note that this does not adjust the progress display in queue, allowing
users to check why their progress never finishes and decide what to do
with the package or links in question.

Fixes #4392