[UnRar] Allow extraction ignoring file attributes (#4395)
This adds a new (default off) option which if enabled passes `-ai` to
`unrar`, which ignores file attributes stored in RAR archives.

This should help with cases where file permissions are set to odd values
that may result in users not being able to modify extracted files. By
ignoring attributes, permissions are handled by the underlying OS
functions of creating a file.